Move container to components folder
