+++ client/components/navigation-container/style.scss
@@ -1,0 +1,23 @@
+.woocommerce-navigation {
+	position: fixed;
+	top: $admin-bar-height;
+	left: 0;
+	width: $navigation-width;
+	height: 100%;
+	box-sizing: border-box;
+
+	.components-navigation {
+		height: 100%;
+		box-sizing: border-box;
+	}
+
+	.components-navigation__level {
+		display: flex;
+		flex-direction: column;
+		height: 100%;
+	}
+
+	.components-navigation__menu:nth-of-type(2) {
+		margin-top: auto;
+	}
+}
