+++ client/index.js
@@ -7,7 +7,7 @@ import { render } from '@wordpress/element';
  * Internal dependencies
  */
 import './stylesheets/index.scss';
-import NavigationContainer from './navigation-container';
+import NavigationContainer from './components/navigation-container';
 
 const navigationRoot = document.getElementById(
 	'woocommerce-embedded-navigation'
