+++ client/stylesheets/index.scss
@@ -7,30 +7,7 @@
 @import '_variables.scss';
 @import '_mixins.scss';
 @import '_colors.scss';
-
-.woocommerce-navigation {
-	position: fixed;
-	top: $admin-bar-height;
-	left: 0;
-	width: $navigation-width;
-	height: 100%;
-	box-sizing: border-box;
-
-	.components-navigation {
-		height: 100%;
-		box-sizing: border-box;
-	}
-
-	.components-navigation__level {
-		display: flex;
-		flex-direction: column;
-		height: 100%;
-	}
-
-	.components-navigation__menu:nth-of-type(2) {
-		margin-top: auto;
-	}
-}
+@import '../components/navigation-container/style';
 
 .has-woocommerce-navigation {
 	#adminmenuwrap,
